+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+-   Expose new method `OlmMachine.getRoomEventEncryptionInfo`.
+
 # matrix-sdk-crypto-wasm v1.2.2
 
 ## Changes in the WASM bindings